backend/requirements.txt
asgiref==3.8.1
Django==5.2.3
djangorestframework==3.16.0
packaging==25.0
pdf2image==1.17.0
pillow==11.2.1
PyMuPDF==1.26.1
pytesseract==0.3.13
sqlparse==0.5.3
tzdata==2025.2
django-cors-headers>=3.18.0